Replace old optimizer with new
