+++ requirements.txt
@@ -5,4 +5,4 @@ pandas
 ortools
 matplotlib-label-lines
 https://github.com/robbmcleod/pyfastnoisesimd/archive/refs/tags/v0.4.3.zip
-git+https://github.com/mroftei/SionnaTorch.git@v0.2.3
+git+https://github.com/mroftei/SionnaTorch.git@v0.3
